+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This file contains highlights of what changes on each version of the [Free Mobil
 - Breaking change: changed the constructor signature.
 - Breaking change: changed the return type of the `sendMessage()` method.
 - Breaking change: dropped the `sendMessage()` helper function.
-- Breaking change: renamed the `DEFAULT_ENDPOINT` constant to `DEFAULT_ENDPOINT`.
+- Breaking change: renamed the `END_POINT` constant to `DEFAULT_ENDPOINT`.
 - Breaking change: raised the required [Node.js](https://nodejs.org) version.
 - Breaking change: using ES2017 features, like async/await functions.
 - Added the `Client.endPoint` property.